Move code memory so we use pages at a time for thread safety on W^X

+++ src/ast.rs
@@ -382,7 +382,7 @@ impl<'a> AstCompiler<'a> {
                         .compiler
                         .get_function_pointer(pause_function.clone())
                         .unwrap();
-                    let pause_function = self.ir.assign_new(pause_function);
+                    let pause_function = self.ir.assign_new(pause_function as usize);
                     self.ir.call_builtin(
                         pause_function.into(),
                         vec![runtime_pointer_reg.into(), stack_pointer],
@@ -412,7 +412,9 @@ impl<'a> AstCompiler<'a> {
 
                 let runtime_ptr = self.compiler.get_runtime_ptr() as usize;
 
-                let mut code = self.ir.compile(lang, error_fn_pointer, runtime_ptr);
+                let mut code = self
+                    .ir
+                    .compile(lang, error_fn_pointer as usize, runtime_ptr);
 
                 let full_function_name = name
                     .clone()
@@ -567,7 +569,7 @@ impl<'a> AstCompiler<'a> {
                     .find_function("beagle.builtin/allocate")
                     .unwrap();
                 let allocate = self.compiler.get_function_pointer(allocate).unwrap();
-                let allocate = self.ir.assign_new(allocate);
+                let allocate = self.ir.assign_new(allocate as usize);
 
                 let size_reg = self.ir.assign_new(struct_type.size());
                 let stack_pointer = self.ir.get_stack_pointer_imm(0);
@@ -620,7 +622,7 @@ impl<'a> AstCompiler<'a> {
                     .find_function("beagle.builtin/allocate")
                     .unwrap();
                 let allocate = self.compiler.get_function_pointer(allocate).unwrap();
-                let allocate = self.ir.assign_new(allocate);
+                let allocate = self.ir.assign_new(allocate as usize);
 
                 let size_reg = self.ir.assign_new(struct_type.size());
                 let stack_pointer = self.ir.get_stack_pointer_imm(0);
@@ -940,7 +942,7 @@ impl<'a> AstCompiler<'a> {
                     .find_function("beagle.builtin/allocate_float")
                     .unwrap();
                 let allocate = self.compiler.get_function_pointer(allocate).unwrap();
-                let allocate = self.ir.assign_new(allocate);
+                let allocate = self.ir.assign_new(allocate as usize);
 
                 let runtime_pointer_reg = self.ir.assign_new(self.compiler.get_runtime_ptr());
 
@@ -1027,7 +1029,7 @@ impl<'a> AstCompiler<'a> {
     fn get_function(&mut self, function_name: &str) -> Value {
         let f = self.compiler.find_function(function_name).unwrap();
         let f = self.compiler.get_function_pointer(f).unwrap();
-        let f = self.ir.assign_new(f);
+        let f = self.ir.assign_new(f as usize);
         f.into()
     }
 
@@ -1222,7 +1224,7 @@ impl<'a> AstCompiler<'a> {
             .unwrap();
         let make_closure = self.compiler.get_function_pointer(make_closure).unwrap();
         let make_closure_reg = self.ir.volatile_register();
-        self.ir.assign(make_closure_reg, make_closure);
+        self.ir.assign(make_closure_reg, make_closure as usize);
         let function_pointer_reg = self.ir.volatile_register();
         self.ir
             .assign(function_pointer_reg, Value::RawValue(function_pointer));
@@ -1355,7 +1357,7 @@ impl<'a> AstCompiler<'a> {
             .unwrap_or_else(|| panic!("could not find function {}", arg));
         assert!(function.is_builtin);
         let function = self.compiler.get_function_pointer(function).unwrap();
-        let function = self.ir.assign_new(function);
+        let function = self.ir.assign_new(function as usize);
         let pointer_reg = self.ir.volatile_register();
         let pointer: Value = self.compiler.get_runtime_ptr().into();
         self.ir.assign(pointer_reg, pointer);
